refactor(auth): standardize useUser imports via centralized hook and fix AnalyticsProvider usage
- Add re-export at src/lib/auth/use-user to avoid provider/store divergence\n- Replace all app imports from auth-store with centralized hook\n- Update AnalyticsProvider to use value-returning useUser()\n- This aligns with Providers tree using components/auth/auth-provider
