+++ src/app/client/client-dashboard.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useMemo, useCallback, Suspense, memo, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
-import { useUser } from '@/lib/store/auth-store';
+import { useUser } from '@/lib/auth/use-user';
 import { useQuery } from '@tanstack/react-query';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
