+++ src/components/messages/contact-search.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { useUser } from '@/lib/store/auth-store';
+import { useUser } from '@/lib/auth/use-user';
 import { cn } from '@/lib/utils';
 import { 
   Dialog,
