+++ src/hooks/use-realtime-bookings.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useQueryClient } from '@tanstack/react-query';
 import { createClient } from '@/lib/supabase/client';
-import { useUser } from '@/lib/store/auth-store';
+import { useUser } from '@/lib/auth/use-user';
 import type { RealtimePostgresChangesPayload } from '@supabase/supabase-js';
 
 /**
