+++ src/lib/auth/use-user.ts
@@ -1,0 +1,5 @@
+'use client';
+
+// Centralized re-export so all app code uses the same auth hook
+export { useUser } from '@/components/auth/auth-provider';
+
